Refactored the basic authentication capability I built into a pluggable AuthenticationStrategy interface and a couple of implementations. Implementers are passed a Thrift Client object for which to execute a login sequence, throwing exceptions on failure per standard Thift exceptions.
Two strategies are now implemented: NoAuthenticationStrategyImpl and BasicAuthenticationStrategyImpl. NoAuthenticationStrategyImpl does nothing, which is intended as the default behavior instead of relying on exceptions or null checks to prevent authentication from being attempted. BasicAuthenticationStrategyImpl is instantiated with a String username and password and when called, executes a login call on the provided Thift Client.

ConnectionPoolConfiguration now requires implementations to offer a getter to return the authentication strategy. ConnectionPoolConfigurationImpl returns NoAuthenticationStrategyImpl by default and offers a setter for setting others.