Reverse iteration no longer causes an infinite loop
The query implementation did not handle reverse iteration correctly. It assumed it was always going forwards and had no way of stopping when going backwards, instead it repeated the first column over and over again. It also did not handle duplicates caused by the reverse iteration starting on the last column of the previous page.

This is not an ideal implementation, but it works for the test case I've set up. The code should probably be rewritten with more robust page handling code both for the forward and reverse cases.