Query the cluster partitioner at startup.
This reverts the previous strategy on this branch which was to make
users configure the partitioner explicitly via a method call to
AstyanaxConfiguration.setPartitioner().  The new approach is to query
Cassandra for the partitioner class the first time the partitioner is
used and cache the result ever after.

There's a chicken-and-egg problem where the connection pool must choose
tokens to identify the seed hosts before any Cassandra queries can be
made.  The code assumes RandomPartitioner and uses BigIntegerToken for
seeds, and as long as all the seeds are replaced at once the first time
host discovery executes then TokenPartitionedTopology never mixes the
initial faked tokens with the actual tokens.